res changed csFrustum to use size_ts for vertex indices. Also made   use of csRefCount for refcouting and use cs_malloc() for internal   vertex array allocation.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@26237 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csgeom/math3d.cpp
@@ -877,10 +877,10 @@ bool csIntersect3::BoxFrustum (const csBox3& box, const csFrustum* frustum)
   csVector3 d = box.Max ()-m;		// Half-diagonal.
 
   const csVector3& origin = frustum->GetOrigin ();
-  int vtcount = frustum->GetVertexCount ();
+  size_t vtcount = frustum->GetVertexCount ();
   csVector3* vts = frustum->GetVertices ();
-  int i1 = vtcount-1;
-  int i;
+  size_t i1 = vtcount-1;
+  size_t i;
   for (i = 0 ; i < vtcount ; i++)
   {
     csPlane3 p (origin, vts[i], vts[i1]);
